+++ app/controllers/handoffs_controller.rb
@@ -7,63 +7,29 @@ class HandoffsController < ApplicationController
   before_action :set_editable_subject
   before_action :redirect_without_subject
 
-  before_action :set_handoff, only: [:show, :edit, :update, :destroy]
-
-  # # GET /handoffs
-  # def index
-  #   @handoffs = Handoff.all
-  # end
-
-  # # GET /handoffs/1
-  # def show
-  # end
-
   # GET /handoffs/new
   def new
-    @handoff = @project.handoffs.new
   end
-
-  # # GET /handoffs/1/edit
-  # def edit
-  # end
 
   # POST /handoffs
   def create
-    @handoff = @project.handoffs.where(user_id: current_user.id, subject_event_id: params[:subject_event_id]).first_or_create
+    @handoff = @project.handoffs.where(handoff_params).first_or_create(user_id: current_user.id)
+    @handoff.set_token
     sign_out @user
     redirect_to handoff_start_path(@project, @handoff)
   end
 
-  # # PATCH /handoffs/1
-  # def update
-  #   if @handoff.update(handoff_params)
-  #     redirect_to @handoff, notice: 'Handoff was successfully updated.'
-  #   else
-  #     render :edit
-  #   end
-  # end
-
-  # # DELETE /handoffs/1
-  # def destroy
-  #   @handoff.destroy
-  #   redirect_to handoffs_path, notice: 'Handoff was successfully destroyed.'
-  # end
-
   private
 
   def set_editable_subject
-    @subject = current_user.all_subjects.find_by_id(params[:id])
+    @subject = current_user.all_subjects.find_by_id params[:id]
   end
 
   def redirect_without_subject
     empty_response_or_root_path(project_subjects_path(@project)) unless @subject
   end
 
-  def set_handoff
-    @handoff = @project.handoffs.find_by_id params[:handoff_id]
-  end
-
   def handoff_params
-    params.require(:handoff).permit(:event_id)
+    { subject_event_id: params[:subject_event_id] }
   end
 end