+++ app/controllers/subjects_controller.rb
@@ -1,13 +1,11 @@
 class SubjectsController < ApplicationController
   before_action :authenticate_user!
-  before_action :set_viewable_project,                  only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report]
-  before_action :set_editable_project_or_editable_site, only: [:new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
-  before_action :redirect_without_project,              only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :sheets, :event, :report, :new, :edit, :create, :update, :destroy, :search, :choose_site, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event]
-  before_action :set_viewable_subject,                  only: [:show, :timeline, :comments, :settings, :files, :adverse_events, :sheets, :event]
-  before_action :set_editable_subject,                  only: [:edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event]
-  before_action :redirect_without_subject,              only: [:show, :timeline, :comments, :settings, :files, :adverse_events, :sheets, :event, :edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :events, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :set_viewable_project,                  only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report, :search, :choose_site]
+  before_action :set_editable_project_or_editable_site, only: [:new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :redirect_without_project,              only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report, :search, :choose_site, :new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :set_viewable_subject,                  only: [:show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event]
+  before_action :set_editable_subject,                  only: [:edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
   before_action :set_design,                            only: [:new_data_entry]
-  before_action :redirect_without_design,               only: [:new_data_entry]
 
   def data_entry
   end
@@ -111,14 +109,23 @@ class SubjectsController < ApplicationController
   ## Find or create subject for the purpose of filling out a sheet for the subject.
   def choose_site
     @subject = current_user.all_viewable_subjects.where(project_id: @project.id).where("LOWER(subjects.subject_code) = ?", params[:subject_code].to_s.downcase).first
-    redirect_to [@project, @subject] if @subject
+    if !@project.site_or_project_editor?(current_user) && !@subject
+      alert_text = params[:subject_code].blank? ? nil : "Subject <code>#{params[:subject_code]}</code> was not found."
+      redirect_to @project, alert: alert_text
+    else
+      redirect_to [@project, @subject] if @subject
+    end
   end
 
   def search
     @subjects = current_user.all_viewable_subjects.where(project_id: @project.id).search(params[:q]).order('subject_code').limit(10)
 
     if @subjects.count == 0
-      render json: [{ value: params[:q], subject_code: params[:q], status_class: 'warning', status: 'NEW' }]
+      if @project.site_or_project_editor? current_user
+        render json: [{ value: params[:q], subject_code: params[:q], status_class: 'warning', status: 'NEW' }]
+      else
+        render json: []
+      end
     else
       render json: @subjects.collect{ |s| { value: s.subject_code, subject_code: s.subject_code, status_class: (s.status == 'valid' ? 'success' : 'info'), status: s.status.first  } }
     end
@@ -210,10 +217,12 @@ class SubjectsController < ApplicationController
 
   def set_viewable_subject
     @subject = current_user.all_viewable_subjects.find_by_id(params[:id])
+    redirect_without_subject
   end
 
   def set_editable_subject
     @subject = current_user.all_subjects.find_by_id(params[:id])
+    redirect_without_subject
   end
 
   def redirect_without_subject
@@ -222,20 +231,14 @@ class SubjectsController < ApplicationController
 
   def set_design
     @design = current_user.all_viewable_designs.where(project_id: @project.id).find_by_param params[:design_id]
-  end
-
-  def redirect_without_design
     empty_response_or_root_path(data_entry_project_subject_path(@project, @subject)) unless @design
   end
 
   def subject_params
     params[:subject] ||= {}
-
     params[:subject][:subject_code] = params[:subject][:subject_code].strip unless params[:subject][:subject_code].blank?
-
     params[:subject][:site_id] = (current_user.all_editable_sites.pluck(:id).include?(params[:site_id].to_i) ? params[:site_id].to_i : nil)
     params[:subject][:project_id] = @project.id
-
     params.require(:subject).permit(
       :project_id, :subject_code, :site_id, :acrostic, :email, :status
     )